Implemented Linear Search Algorithm in C++
LinearSearch is a method which takes 3 arguments namely array[] in which iteration will be done, length is the length of the array, key is the number which needs to be found in the array.  The method return an integer. If the number given is found in the array, the index of the number will be returned else it will return -1 which means that the number given as the key was not found in the given array.